+++ tests/testthat/helper.R
@@ -1,0 +1,3 @@
+has_ws <- function(x) {
+  any(!is.na(x) & (startsWith(x, " ") | endsWith(x, " ")))
+}